refactor auth files
